requirements.txt
asgiref==3.6.0
autopep8==2.0.1
beautifulsoup4==4.11.2
certifi==2022.12.7
charset-normalizer==3.1.0
debugpy==1.6.6
distlib==0.3.7
Django==3.2.16
django-bootstrap-v5==1.0.11
django-bootstrap5==22.2
django-environ==0.9.0
django-filter==22.1
django-recaptcha==3.0.0
djangorestframework==3.14.0
docopt==0.6.2
filelock==3.12.2
idna==3.4
importlib-metadata==6.7.0
Jinja2==3.1.2
Markdown==3.4.1
MarkupSafe==2.1.2
mysqlclient==2.1.1
pipenv==2023.9.8
pkg-resources==0.0.0
platformdirs==3.10.0
pycodestyle==2.10.0
PyMySQL==1.0.2
pytz==2022.7
redmail==0.6.0
requests==2.28.2
soupsieve==2.3.2.post1
sqlparse==0.4.3
tomli==2.0.1
typing-extensions==4.7.1
urllib3==1.26.15
virtualenv==20.24.5
yarg==0.1.9
zipp==3.11.0
